product type initializer added

+++ src/main/java/com/GraduationProject/GraduationProject/dto/ProductTypeDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ProductTypeDTO {
     private String productName;
-    private Float effect;
+    private Double effect;
